+++ lib/src/state/scribble.state.dart
@@ -1,6 +1,5 @@
 import 'dart:ui';
 
-import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:scribble/src/model/sketch/sketch.dart';